Improve scrolling behaviour (#2)
* Keep cursor still when scrolling off-screen with keyboard

Now using PageUp and PageDown scroll exactly one page, and will keep the
cursor in the same page on the screen, instead of centering it. In the
same vein, when using the arrow keys to move the cursor off-screen, the
view will now only be scrolled by a single cell instead of jumping to
move the cursor to the center.

* Prevent dragging off-screen causing wild jumps

* Scroll smoothly when dragging outside the grid