Add exporting as a nexus script
Calculating the diagnostics does most of the work of finding double taps
and figuring out whether the dash key is held, so I have just tacked the
nexus script generation onto the end of that. This still doesn't try to
handle cases where using left+right (or up+down) to make neutral inputs
would be needed.